+++ Form1.cs
@@ -173,7 +173,7 @@ namespace uclliu
             }
             //如果是需要跳過的 app ，就跳過
             var p_info = ucl.getForegroundWindowProcessInfo();
-            if (my.in_array(p_info["PROCESS_NAME"].ToString(), ucl.sendkey_not_use_ucl_apps))
+            if (TextOutputRouter.MatchesProcess(p_info["PROCESS_NAME"].ToString(), ucl.sendkey_not_use_ucl_apps))
             {
                 if (ucl.is_ucl())
                 {